+++ src/main/java/springredis/demo/Service/impl/FuzzyStringMatchImpl.java
@@ -1,4 +1,4 @@
-package springredis.demo.Service;
+package springredis.demo.Service.impl;
 
 
 import springredis.demo.structures.FuzzyStringMatch.EditOp;
