+++ src/main/java/springredis/demo/entity/JourneyJsonModel.java
@@ -17,7 +17,7 @@ public class JourneyJsonModel {
         String updatedBy;
         String journeyId;
         String thumbNailURL;
-        int status;
+        Integer status;
         String stage;
     }
 }
